[OrderedSet] Split SetAlgebra predicates into individual files
This is to match the layout of all the other
SetAlgebra implementations in this package.
